+++ apig-20240327/src/Models/CreateMcpServerRequest.php
@@ -23,6 +23,11 @@ class CreateMcpServerRequest extends Model
     /**
      * @var string
      */
+    public $createFromType;
+
+    /**
+     * @var string
+     */
     public $description;
 
     /**
@@ -67,6 +72,7 @@ class CreateMcpServerRequest extends Model
     protected $_name = [
         'assembledSources' => 'assembledSources',
         'backendConfig' => 'backendConfig',
+        'createFromType' => 'createFromType',
         'description' => 'description',
         'domainIds' => 'domainIds',
         'exposedUriPath' => 'exposedUriPath',
@@ -111,6 +117,10 @@ class CreateMcpServerRequest extends Model
 
         if (null !== $this->backendConfig) {
             $res['backendConfig'] = null !== $this->backendConfig ? $this->backendConfig->toArray($noStream) : $this->backendConfig;
+        }
+
+        if (null !== $this->createFromType) {
+            $res['createFromType'] = $this->createFromType;
         }
 
         if (null !== $this->description) {
@@ -180,6 +190,10 @@ class CreateMcpServerRequest extends Model
 
         if (isset($map['backendConfig'])) {
             $model->backendConfig = backendConfig::fromMap($map['backendConfig']);
+        }
+
+        if (isset($map['createFromType'])) {
+            $model->createFromType = $map['createFromType'];
         }
 
         if (isset($map['description'])) {